Rename funtion_pipe -> metafunctions
