+++ metafunctions/tests/test_function_chain.py
@@ -1,9 +1,9 @@
 import operator
 
-from function_pipe.metafunctions import FunctionChain
-from function_pipe.metafunctions import FunctionMerge
-from function_pipe.metafunctions import SimpleFunction
-from function_pipe.tests.util import BaseTestCase
+from metafunctions.metafunctions import FunctionChain
+from metafunctions.metafunctions import FunctionMerge
+from metafunctions.metafunctions import SimpleFunction
+from metafunctions.tests.util import BaseTestCase
 
 
 class TestUnit(BaseTestCase):
