+++ metafunctions/tests/test_function_merge.py
@@ -1,8 +1,8 @@
 import operator
 
-from function_pipe.metafunctions import FunctionMerge
-from function_pipe.metafunctions import SimpleFunction
-from function_pipe.tests.util import BaseTestCase
+from metafunctions.metafunctions import FunctionMerge
+from metafunctions.metafunctions import SimpleFunction
+from metafunctions.tests.util import BaseTestCase
 
 
 class TestUnit(BaseTestCase):
